⚡️  (next) dynamic !ssr is not ... well working
These all seem to be improvements that are more trouble than they are worth
