+++ sites/jeromefitzgerald.com/src/components/NowPlaying/NowPlaying.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { Separator } from '@jeromefitz/ds/components/Separator'
 import { Suspense } from 'react'
