+++ sites/jeromefitzgerald.com/src/components/NowReading/NowReading.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { Separator } from '@jeromefitz/ds/components/Separator'
 import { cx } from '@jeromefitz/ds/utils/cx'
