+++ sites/jeromefitzgerald.com/src/components/Playground/index.ts
@@ -1,3 +1,4 @@
+export { PlaygroundPage } from './Playground.Page'
 export { RadixBackgrounds } from './Radix.Backgrounds'
 export { RadixBorders } from './Radix.Borders'
 export { RadixColors } from './Radix.Colors'
